requirements.txt
pyArubaCloud==0.7
click